+++ src/mailer/dto/create-mailer.dto.ts
@@ -1,0 +1,1 @@
+export class CreateMailerDto {}
