+++ src/mailer/dto/update-mailer.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateMailerDto } from './create-mailer.dto';
+
+export class UpdateMailerDto extends PartialType(CreateMailerDto) {}
