+++ src/mailer/entities/mailer.entity.ts
@@ -1,0 +1,1 @@
+export class Mailer {}
